5/3
몬스터 스폰 로직 구현완

+++ 2d/Assets/Project/Scripts/Gameplay/Enemy/Drop/ExpManager.cs
@@ -55,7 +55,7 @@ public class ExpManager : MonoBehaviour
 
 
     // 적이 죽을 때 호출할 함수
-    public void DropExp(Vector2 position,int expAmount)
+    public void DropExp(Vector2 position,float expAmount)
     {
         Exp exp = pool.Get();
         exp.SetExp(expAmount);
